Make more fixes and improvements
- Remove PAGESIZE constant
- Fix realloc() documentation
- Fix ttyname_r() error reporting
- Make forking more reliable on Windows
- Make execvp() a few microseconds faster
- Make system() a few microseconds faster
- Tighten up the socket-related magic numbers
- Loosen restrictions on mmap() offset alignment
- Improve GetProgramExecutableName() with getenv("_")
- Use mkstemp() as basis for mktemp(), tmpfile(), tmpfd()
- Fix flakes in pthread_cancel_test, unix_test, fork_test
- Fix recently introduced futex stack overflow regression
- Let sockets be passed as stdio to subprocesses on Windows
- Improve security of bind() on Windows w/ SO_EXCLUSIVEADDRUSE

+++ libc/calls/read.c
@@ -86,7 +86,7 @@ ssize_t read(int fd, void *buf, size_t size) {
     rc = ebadf();
   }
   END_CANCELLATION_POINT;
-  DATATRACE("read(%d, [%#.*hhs%s], %'zu) → %'zd% m", fd, MAX(0, MIN(40, rc)),
-            buf, rc > 40 ? "..." : "", size, rc);
+  DATATRACE("read(%d, [%#.*hhs%s], %'zu) → %'zd% m", fd,
+            (int)MAX(0, MIN(40, rc)), buf, rc > 40 ? "..." : "", size, rc);
   return rc;
 }
